rename mock sleep functions
